+++ hatch_conda/hooks.py
@@ -1,6 +1,3 @@
-# SPDX-FileCopyrightText: 2021-present Ofek Lev <oss@ofek.dev>
-#
-# SPDX-License-Identifier: MIT
 from hatchling.plugin import hookimpl
 
 from .plugin import CondaEnvironment