+++ HISTORY.md
@@ -6,7 +6,22 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 12/08/2022
 
-## 0.1.0 - 2022-08-12
+### Added 
 
-This is the initial public release.
+ - Trove classifier for hatch framework.
+ - Option to use mamba.
+ - Option to disable `conda-forge` index.
+
+### Fixed
+
+ - Fixed entry point name.
+ - Unneccesary `Activate` and `Deactivate` calls removed.
+
+
+## [0.1.0] - 12/08/2022
+
+### Added
+
+ - Conda environment type for hatch.